fix changes of MapElement in other files that use it

+++ src/main/java/it/unibo/isaccoop/model/room/RoomFactory.java
@@ -13,36 +13,36 @@ public interface RoomFactory {
      * @param coordInsideLevel the coordinate of this room inside the level
      * @return an empty room where the protagonist starts the game.
      */
-    Room buildStartRoom(Pair<Integer, Integer> coordInsideLevel);
+    Room buildStartRoom(Pair<Double, Double> coordInsideLevel);
 
     /** 
      * @param coordInsideLevel the coordinate of this room inside the level
      * @return a room with enemies, obstacles and doors.
      */
-    Room buildStandardRoom(Pair<Integer, Integer> coordInsideLevel);
+    Room buildStandardRoom(Pair<Double, Double> coordInsideLevel);
 
     /** 
      * @param coordInsideLevel the coordinate of this room inside the level
      * @return a room where you can ONLY buy powerups using the collected money.
      */
-    Room buildShopRoom(Pair<Integer, Integer> coordInsideLevel);
+    Room buildShopRoom(Pair<Double, Double> coordInsideLevel);
 
     /** 
      * @param coordInsideLevel the coordinate of this room inside the level
      * @return the room with the final boss
      */
-    Room buildBossRoom(Pair<Integer, Integer> coordInsideLevel);
+    Room buildBossRoom(Pair<Double, Double> coordInsideLevel);
 
     /** 
      * @param coordInsideLevel the coordinate of this room inside the level
      * @return a room where you can find a free powerup
      */
-    Room buildTreasureRoom(Pair<Integer, Integer> coordInsideLevel);
+    Room buildTreasureRoom(Pair<Double, Double> coordInsideLevel);
 
     /** 
      * @param roomType type of the room to be created
      * @param coordInsideLevel the coordinate of this room inside the level
      * @return a room where you can find a free powerup
      */
-    Room buildRoomOfType(RoomType roomType, Pair<Integer, Integer> coordInsideLevel);
+    Room buildRoomOfType(RoomType roomType, Pair<Double, Double> coordInsideLevel);
 }
